Урок 7

+++ Lesson7/Account.java
@@ -1,16 +1,16 @@
 package Lesson7;
 
 
-
 public class Account {
 
-    private Integer id ;
-    private String holder ;
-    private double amount = 0;
+    private int id;
+    private String holder;
+    private int amount = 0;
 
 
-    public  Account(Integer id, String holder) {
-    this.id = id;
-    this.holder = holder; }
+    public Account(Integer id, String holder) {
+        this.id = id;
+        this.holder = holder;
+    }
 
 }
